requirements.txt
django==4.2.1
djangorestframework==3.15.2
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pillow==10.4.0
python-dotenv==1.0.1